Create expression_classifier.py
An expression classifier that can be trained on pictures of facial expressions to classify expressions as "Sad, Happy, Angry, or Neutral" using a Keras Neural Network. Then pickles the file automatically. Pickled file can then be loaded into the other program to analyze and help matplotlib visualize the data for you.